[jit] fix compilation order for class methods
It turns out that methods are reported in Python in name order instead
of definition order. This means that potentially we can try to compile a
method before __init__, which will break if we wanted to use any
attributes that get defined in __init__.

This PR fixes that by special-casing __init__ so that its always
compiled first. Fixes #20004.

gh-metadata: pytorch pytorch 20094 gh/suo/34/head